MWPW-192158: Rename the translation-items-loader file to items-loader
