+++ studio/test/translation/translation-items-loader.test.js
@@ -13,7 +13,7 @@ import {
     fetchUnresolvedVariations,
     fetchVariationByPath,
     setCardVariationsByPaths,
-} from '../../src/common/utils/translation-items-loader.js';
+} from '../../src/common/utils/items-loader.js';
 
 describe('translation-items-loader', () => {
     let sandbox;
